Re #26955 Remove SANS Scale workflow alg.
Removes SANS Scale Workflow algorithm and turns it into a Python class
