+++ Framework/PythonInterface/test/python/plugins/algorithms/WorkflowAlgorithms/sans/CMakeLists.txt
@@ -7,7 +7,6 @@ set(TEST_PY_FILES
     SANSCreateAdjustmentWorkspacesTest.py
     SANSCreateWavelengthAndPixelAdjustmentTest.py
     SANSNormalizeToMonitorTest.py
-    SANSScaleTest.py
     SANSSliceEventTest.py)
 
 check_tests_valid(${CMAKE_CURRENT_SOURCE_DIR} ${TEST_PY_FILES})
