+++ scripts/test/SANS/algorithm_detail/CMakeLists.txt
@@ -7,6 +7,7 @@ set(TEST_PY_FILES
     move_workspaces_test.py
     q_resolution_calculator_test.py
     scale_helper_test.py
+    scale_sans_workspace_test.py
     strip_end_nans_test.py
     centre_finder_new_test.py
     batch_execution_test.py)
